feat: remove unused context, refactor pool structure
