+++ views/pool/components/StakedPoolCardItem/index.tsx
