+++ views/pool/components/StakedPoolListItem/index.tsx
